Auto registration as CAP plugin
- The middleware registers itself when loaded as a CAP plugin
- Declarative configuration through `package.json` / `cds-rc.json`
- Programmatic registration is still possible, but disables auto mode then
