+++ lib/etc.js
@@ -1,0 +1,1 @@
+module.exports.registered = Symbol.for('swaggerui_registered')
