+++ package.json
@@ -24,6 +24,8 @@
   },
   "files": [
     "CHANGELOG.md",
+    "lib/",
+    "cds-plugin.js",
     "index.js"
   ],
   "dependencies": {
@@ -49,5 +51,48 @@
   "jest": {
     "testEnvironment": "node",
     "testTimeout": 10000
+  },
+  "cds": {
+    "schema": {
+      "cds": {
+        "swagger": {
+          "oneOf": [
+            {
+              "type": "object",
+              "properties": {
+                "basePath": {
+                  "type": "string",
+                  "description": "The root path to mount the middleware on",
+                  "default": "/$api-docs"
+                },
+                "apiPath": {
+                  "type": "string",
+                  "description": "The root path for the services (useful if behind a reverse proxy)"
+                },
+                "diagram": {
+                  "type": "boolean",
+                  "description": "Whether to render the YUML diagram",
+                  "default": true
+                },
+                "odataVersion": {
+                  "type": "string",
+                  "description": "the OData Version to compile the OpenAPI specs"
+                }
+              }
+            },
+            {
+              "type": "boolean",
+              "description": "Shortcut to disable Swagger UI"
+            }
+          ]
+        }
+      }
+    },
+    "swagger": {
+      "basePath": "/$api-docs",
+      "apiPath": "",
+      "diagram": true,
+      "odataVersion": ""
+    }
   }
 }
