+++ tests/app/.npmrc
@@ -1,0 +1,1 @@
+package-lock=false
