+++ tests/app/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "test-app",
+  "version": "1.0.0",
+  "devDependencies": {
+    "cds-swagger-ui-express": "file:../.."
+  },
+  "cds": {
+    "swagger": {
+      "basePath": "/$api-docs-plugin"
+    }
+   }
+}
