+++ tests/server.test.js
@@ -1,7 +1,8 @@
-const cds = require('@sap/cds/lib')
-const { GET, expect } = cds.test.in(__dirname, 'app').run('serve', 'all')
+const cds = require('@sap/cds')
 
-describe('Swagger UI', ()=>{
+describe.only('Swagger UI', ()=>{
+
+  const { GET, expect } = cds.test.in(__dirname, 'app').run('serve', 'all')
 
   test('Service endpoint', async()=>{
     const { data } = await GET `/browse/$metadata`
@@ -27,6 +28,20 @@ describe('Swagger UI', ()=>{
 
     data  = (await GET `/$api-docs/admin/swagger-ui-init.js`).data
     expect (data ) .to.be.a('string').that.contains('AdminService')
+  })
+
+})
+
+describe('Swagger UI as plugin', ()=>{
+
+  beforeAll(()=> process.env.TEST_AS_PLUGIN = true)
+  afterAll(()=> delete process.env.TEST_AS_PLUGIN)
+
+  const { GET, expect } = cds.test.in(__dirname, 'app').run('serve', 'all')
+
+  test('Main HTML', async()=>{
+    const { data } = await GET `/$api-docs-plugin/browse`
+    expect (data) .match (/swagger/i)
   })
 
 })
